+++ ll/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added a new W5500 type for VDM transfers with an infallible GPIO and an infallible SPI bus.
+- Added `defmt::Format` implementations for all register structs.
+- Added the missing `opmd` field to the `Display` impl for `PhyCfg`.
+
+## Fixed
+- Fixed the `opmdc` field name in the `Display` impl for `PhyCfg`.  This field was incorrectly named `opmd`.
 
 ## [0.10.3] - 2022-04-27
 ### Added